cambios en el tipo de dato del campo telefono

+++ database/migrations/2024_11_12_054047_create_coordinadors_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name')->nullable();
             $table->string('ap_paterno')->nullable();
             $table->string('ap_materno')->nullable();
-            $table->integer('telefono')->nullable();
+            $table->bigInteger('telefono')->nullable();
             $table->string('correo')->nullable();
             $table->timestamps();
         });
